Make original detectron2 and onnx separate models
