+++ unstructured_inference/models/base.py
@@ -5,6 +5,10 @@ from unstructured_inference.models.detectron2 import (
     MODEL_TYPES as DETECTRON2_MODEL_TYPES,
     UnstructuredDetectronModel,
 )
+from unstructured_inference.models.detectron2onnx import (
+    MODEL_TYPES as DETECTRON2_ONNX_MODEL_TYPES,
+    UnstructuredDetectronONNXModel,
+)
 from unstructured_inference.models.yolox import (
     MODEL_TYPES as YOLOX_MODEL_TYPES,
     UnstructuredYoloXModel,
@@ -18,6 +22,9 @@ def get_model(model_name: Optional[str] = None) -> UnstructuredModel:
     if model_name in DETECTRON2_MODEL_TYPES:
         model: UnstructuredModel = UnstructuredDetectronModel()
         model.initialize(**DETECTRON2_MODEL_TYPES[model_name])
+    elif model_name in DETECTRON2_ONNX_MODEL_TYPES:
+        model = UnstructuredDetectronONNXModel()
+        model.initialize(**DETECTRON2_ONNX_MODEL_TYPES[model_name])
     elif model_name in YOLOX_MODEL_TYPES:
         model = UnstructuredYoloXModel()
         model.initialize(**YOLOX_MODEL_TYPES[model_name])
